+++ tapo-py/src/handlers/child_devices.rs
@@ -1,3 +1,5 @@
+mod t110_handler;
 mod t31x_handler;
 
+pub use t110_handler::*;
 pub use t31x_handler::*;